requirements.txt
python-dotenv
langchain-mcp-adapters
langgraph
langchain[openai]
mcp
streamlit
